Create README.md
🚀 Exciting Update! 🎉

I’m thrilled to share that I have successfully completed the Introduction to Career Skills in Data Analytics course on LinkedIn Learning! 📊

This course has provided me with valuable insights into the world of data analytics, covering essential skills and strategies to navigate my career in this dynamic field. I’m eager to apply what I’ve learned and take on new challenges in data analysis.

A big thank you to the instructors for their guidance and to LinkedIn Learning for this opportunity!

#DataAnalytics #CareerDevelopment #ContinuousLearning #LinkedInLearning